+++ hosts/phi/configuration.nix
@@ -9,9 +9,6 @@
 
   hardware.cpu.amd.updateMicrocode = true;
 
-  # Living on the edge for Navi10
-  boot.kernelPackages = pkgs.linuxPackages_latest;
-
   networking.nameservers = [ "1.1.1.1" ];
   networking.dhcpcd.extraConfig = ''
     nohook resolv.conf