+++ modules/greetd.nix
@@ -3,6 +3,10 @@
     services.greetd.enable = true;
     programs.regreet.enable = true;
 
+    programs.regreet.font.name = "DejaVu Sans";
+    programs.regreet.font.size = 12;
+    programs.regreet.font.package = pkgs.dejavu_fonts;
+
     services.greetd.settings.default_session.command =
       "${lib.getExe' pkgs.dbus "dbus-run-session"} ${
         lib.getExe pkgs.sway