+++ modules/pim.nix
@@ -1,8 +1,7 @@
 {
   nixosModule = { user, pkgs, ... }: {
     environment.systemPackages = builtins.attrValues {
-      inherit (pkgs) gnome-calendar gnome-contacts;
-      inherit (pkgs.gnome) gnome-control-center;
+      inherit (pkgs) gnome-calendar gnome-contacts gnome-control-center;
     };
 
     services.gnome.gnome-online-accounts.enable = true;